ImageSearch: Add multi-service IQDB search to available services
Also refactored the way available search services are retrieved. All of the services are now contained in SearchServices class as one collection, and MainView binds that to the combo box's items source. This eliminates unnecessary work if we were to add additional services, since there'd be only one file we'll have to edit.