ImageSearch: Replace SearchServiceView with item template binding
The issue with service locator is when it displays transition between two views. During the fade out the view was stretching horizontally to fit the container combo box, which in turn increased its height and made the UI jump up and down.

While it may not be the best solution, it is better than overriding default template for the view host locator to disable the fade effect. Maybe in the future.