BooruDownloader: Handle selected items properly
Now observable collection and OWB from view model to view (which have always been the correct way to do selections) are used instead of hacked enumerable and OWB from view to view model.

In the end the final result is still hacked because WPF's ListBox doesn't care if its SelectedItems is bound to an observable collection.